tablon_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiomysql==0.2.0
SQLAlchemy==2.0.30
alembic==1.13.1
python-dotenv==1.0.1
pydantic[email]==2.7.1
websockets==12.0
